requirements.txt
django==4.2
Pillow  # manejo de imagenes
mysqlclient
django-widget-tweaks
requests
python-decouple
sudo apt install libpq-dev
pip install psycopg